Attempt to remove dirty workarounds for race condition upon stopping ErrorFilter, and solve the problem in a proper way.
ErrorFilter is a forked process. When it gets killed at the end of 
the test, it destroys connections to servers. If after the test 
we check replication status, wait for slave to catch up with master, 
etc., we have to reconnect to the server. Since killing ErrorFilter
is performed concurrently with other activities, it can happened at 
any moment, so we need to handle the sudden loss of connection in 
various places. 

The right way to solve it is to make sure that killing ErrorFilter
does not affect connections created before it was forked. 
InactiveDestroy is supposed to do just that, although it will require
extensive testing to make sure it is sufficient in all situations.