Introduce the concept of cleanup block in transformers
Some transformations can end prematurely and leave the environment 
in a dirty state, e.g. with some variables changed.
If a transformation changes the environment, it must have a block marked 
as TRANSFORM_CLEANUP (as a comment in the first statement of the block).
If such a block exists, it will be executed even if the transformation 
is going to quit.