Introduce version compatibility mechanism
Test run can be configured to be compatible with certain
MariaDB/MySQL version. If no configuration option is provided,
compatibility is set to the minimal version of servers used
in the test (not applicable to downgrade tests).

It is not guaranteed that there will be no statements not compatible
with the given version. Only generation mechanisms which were made
aware of this configuration parameter will comply with it