Introduce the concept of "factors" in grammar component definition
For adjusting probabilities of using a component from a rule,
instead of repeating the component multiple times as we did before, e.g.

rule:
  component1 |
  component2 | component2 | component2 | component2 |
  component3 | component3
;

now it is possible to give a numeric factor to a rule.
Semantics is mainly the same -- it defines how many times the component
will be present in the rule.
The difference is that it is also possible to use values in [0,1) range.
A positive value below 1 means that the number of other components will
be multiplied accordingly. For example:

rule:
  ==FACTOR:1== component1 |
  ==FACTOR:4== component2 |
  ==FACTOR:2== component3
;

and

rule:
  ==FACTOR:0.5== component1 |
  ==FACTOR:2==   component2 |
                 component3
;

mean the same, that in the resulting rule component1 will be present
once, component3 -- twice, and component2 -- four times.

The second variant is useful when components are added gradually,
so that all factors don't need to be re-written.

0 will mean that the rule won't be used (not very practical, but let it be).