Work around "too many program headers" problem in the coredump
When it occurs (apparently if too many plugins are used at once),
file <coredump> does not show the path to the binary, so we will
rely on the one stored in the server definition. It can be wrong,
e.g. if the server binary was different at the time of the
coredump-ing, or if it's not from the server at all; but that's
the "best effort" attempt